Commencer #6

+++ JeuEchecs/src/com/echecs/pieces/Cavalier.java
@@ -11,7 +11,7 @@ public class Cavalier extends Piece
     public boolean peutSeDeplacer(Position pos1, Position pos2, Piece[][] echiquier)
     {
         //Deplacement en X et Y
-        int depX = Math.abs(pos2.getColonne() - pos1.getColonne());
+        int depX = Math.abs((int)pos2.getColonne() - (int)pos1.getColonne());
         int depY = Math.abs(pos2.getLigne() - pos1.getLigne());
 
         return depX == 1 && depY == 2 || depX == 2 && depY == 1;
